getinfo.shape {maptools}	R Documentation

Get shapefile header information

Description

Get shapefile header information; the file should be given including its ".shp" extension, and the function will reconstruct the names of the database (dbf) file and the index (shx) file from these.

Usage

getinfo.shape(filen)
## S3 method for class 'shapehead'
print(x, ...)

Arguments

filen
name of file with *.shp extension

x
a shapehead list as returned by getinfo.shape

...
other arguments passed to print

Details

The function calls code from shapelib to read shapefiles, a file format used by ESRI GIS software among others

Value

The function returns a list of class shapehead.

Author(s)

Roger Bivand Roger.Bivand@nhh.no; shapelib by Frank Warmerdam

References

http://shapelib.maptools.org/

Examples

res <- getinfo.shape(system.file("shapes/fylk-val.shp", package="maptools")[1])
res
str(res)

[Package maptools version 1.1-4 Index]