spCbind-methods {maptools}	R Documentation

cbind for spatial objects

Description

spCbind provides cbind-like methods for Spatial*DataFrame objects in addition to the $, [<- and [[<- methods already available.

Methods

obj = "SpatialPointsDataFrame", x = "data.frame"

cbind a data frame to the data slot of a SpatialPointsDataFrame object

obj = "SpatialPointsDataFrame", x = "vector"

cbind a vector to the data slot of a SpatialPointsDataFrame object

obj = "SpatialLinesDataFrame", x = "data.frame"

cbind a data frame to the data slot of a SpatialLinesDataFrame object; the data frame argument must have row names set to the Lines ID values, and should be re-ordered first by matching against a shared key column

obj = "SpatialLinesDataFrame", x = "vector"

cbind a vector to the data slot of a SpatialLinesDataFrame object

obj = "SpatialPolygonsDataFrame", x = "data.frame"

cbind a data frame to the data slot of a SpatialPolygonsDataFrame object; the data frame argument must have row names set to the Polygons ID values, and should be re-ordered first by matching against a shared key column

obj = "SpatialPolygonsDataFrame", x = "vector"

cbind a vector to the data slot of a SpatialPolygonsDataFrame object

Author(s)

Roger Bivand

See Also

spChFIDs-methods, spRbind-methods

Examples

xx <- readShapePoly(system.file("shapes/sids.shp", package="maptools")[1], 
  IDvar="FIPSNO", proj4string=CRS("+proj=longlat +ellps=clrk66"))
xtra <- read.dbf(system.file("share/nc_xtra.dbf", package="maptools")[1])
o <- match(xx$CNTY_ID, xtra$CNTY_ID)
xtra1 <- xtra[o,]
row.names(xtra1) <- xx$FIPSNO
xx1 <- spCbind(xx, xtra1)
names(xx1)
identical(xx1$CNTY_ID, xx1$CNTY_ID.1)

[Package maptools version 0.8-25 Index]