wrld_simpl {maptools}	R Documentation

Simplified world country polygons

Description

The object loaded is a SpatialPolygonsDataFrame object containing a slightly modified version of Bjoern Sandvik's improved version of world_borders.zip - TM_WORLD_BORDERS_SIMPL-0.2.zip dataset from the Mapping Hacks geodata site. The country Polygons objects and the data slot data frame row numbers have been set to the ISO 3166 three letter codes.

Usage

data(wrld_simpl)

Format

The format is: Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots; the data clot contains a data.frame with 246 obs. of 11 variables:

FIPS

factor of FIPS country codes

ISO2

factor of ISO 2 character country codes

ISO3

factor of ISO 3 character country codes

UN

integer vector of UN country codes

NAME

Factor of country names

AREA

integer vector of area values

POP2005

integer vector of population in 2005

REGION

integer vector of region values

SUBREGION

integer vector of subregion values

LON

numeric vector of longitude label points

LAT

numeric vector of latitude label points

The object is in geographical coordinates using the WGS84 datum.

Source

Originally “http://mappinghacks.com/data/TM_WORLD_BORDERS_SIMPL-0.2.zip”, now available from https://github.com/nasa/World-Wind-Java/tree/master/WorldWind/testData/shapefiles

Examples

data(wrld_simpl)
plot(wrld_simpl)

[Package maptools version 1.1-4 Index]