getKMLcoordinates {maptools}	R Documentation

Get a list of coordinates out of a KML file

Description

This function parses a KML file to get the content of <coordinates> tags and returns a list of matrices representing the longitude-latitute or if ignoreAltitude is FALSE the longitude-latitute-altitude coordinates of a KML geometry.

Usage

getKMLcoordinates(kmlfile, ignoreAltitude=FALSE)

Arguments

kmlfile
connection object or a character string of the KML file

ignoreAltitude
if set to TRUE the altitude values of a KML points will be ignored

Value

coords is a list of matrices representing the longitude-latitute or if ignoreAltitude is FALSE the longitude-latitute-altitude coordinates

Author(s)

Hans-J. Bibiko

See Also

kmlPolygon, kmlLine

Examples

data(wrld_simpl)
## creates a KML file containing the polygons of South Africa (plus hole)
sw <- slot(wrld_simpl[wrld_simpl$NAME=="South Africa",], "polygons")[[1]]
tf <- tempfile()
kmlPolygon(sw, kmlfile=tf, name="South Africa", col="#df0000aa", lwd=5, 
    border=4, kmlname="R Test", 
    kmldescription="This is <b>only</b> a <a href='http://www.r-project.org'>R</a> test.")
zz <- getKMLcoordinates(tf, ignoreAltitude=TRUE)
str(zz)
zz <- getKMLcoordinates(system.file("shapes/Testing.kml", package="maptools"))
str(zz)

[Package maptools version 1.1-4 Index]